Correction smeaine 2 Fred Simard.  Principalement orthographique et un peu d'optimisation de code
