+++ src/enginCartes/Carte.java
@@ -6,18 +6,25 @@ package enginCartes;
  * 
  * 
  * Liste des méthodes publiques: 
- *     - Carte, constructeur avec liste vide
- *     - Carte, constructeur à partir de sections de listes
- *     - ajouterLien, méthode permettant d'ajouter un lien
- *     - enleverLien, méthode permettant d'enlever un lien à l'index
+ *     - Carte, constructeur avec liste vide.
+ *     - Carte, constructeur à partir de sections de listes.
+ *     - ajouterLien, méthode permettant d'ajouter un lien.
+ *     - enleverLien, méthode permettant d'enlever un lien à l'index.
  *     - obtientFraction, obtient une fraction de la liste (deep copy)
- *     - getNbLiens, méthode pour obtenir le nombre de liens
- *     - getScore, retourne le score
- *     - evalueScore, calcul le score de la carte
- *     - calculSommeLongueurs, calcul la somme des longeurs
- *     - toString, obtient une représentation chaînes de caractères de l'objet
+ *     - getNbLiens, méthode pour obtenir le nombre de liens..
+ *     - getScore, retourne le score.
+ *     - evalueScore, calcule le score de la carte.
+ *     - calculSommeLongueurs, calcul la somme des longeurs.
+ *     - toString, obtient une représentation chaîne de caractères de l'objet.
  *
- * @author Fred Simard | ETS, 
+ * @author Fred Simard | ETS,
+ *  
+ * @revision Pierre Bélisle | ETS
+ * 
+ *              principalement : Mettre le code sur 80 colonne max.
+ *                               Aération.
+ *                               Correction orthographique. 
+ * 
  * @version Hiver 2018
  */
 
@@ -25,18 +32,22 @@ import listeChainee.ListeDChainee;
 
 public class Carte{
     
-    // liste chainee
+    // Liste chainee des liens.
 	ListeDChainee listeLiens = new ListeDChainee();
+	
 	double score = Double.POSITIVE_INFINITY;
+	
 	MoteurDistanceMoyenne moteurDistanceMoyenne;
 
 	/**
 	 * Constructeur qui reçoit le moteur de distance et initialise avec une
-	 * liste de liens vide
+	 * liste de liens vide.
 	 * 
-	 * @param moteurDistanceMoyenne, reference au moteur de calcul de la distance moyenne
+	 * @param moteurDistanceMoyenne, reference au moteur de calcul de la 
+	 * n                             distance moyenne
 	 */
 	public Carte(MoteurDistanceMoyenne moteurDistanceMoyenne){
+		
 		this.moteurDistanceMoyenne = moteurDistanceMoyenne;
 	}
 
@@ -44,7 +55,8 @@ public class Carte{
 	 * Constructeur qui reçoit le moteur de distance et des segments de listes 
 	 * pour initialiser la liste
 	 * 
-	 * @param moteurDistanceMoyenne, reference au moteur de calcul de distance moyenne
+	 * @param moteurDistanceMoyenne, reference au moteur de calcul de  la 
+	 *                               distance moyenne
 	 * @param section1, premiere section de liste
 	 * @param section2, deuxieme section de liste
 	 */
@@ -52,21 +64,23 @@ public class Carte{
 			     ListeDChainee section1, 
 			     ListeDChainee section2){
 		
-		// copie la référence au moteur de calcul distance
+		// Copie la référence au moteur de calcul distance.
 		this.moteurDistanceMoyenne = moteurDistanceMoyenne;
-		// démarre à partir de la premiere section
+		
+		// Démarre à partir de la premiere section.
 		listeLiens = section1;
 		
 		section2.deplacerDebut();
 		
 		for(int i=0;i<section2.getNbElements();i++){
+			
 			listeLiens.ajouterFin(section2.getElement());
 			section2.deplacerSuivant();
 		}
 	}
 
 	/**
-	 * Méthode permettant d'ajouter un lien à la fin
+	 * Méthode permettant d'ajouter un lien à la fin.
 	 * 
 	 * @param ceLien, lien à ajouter
 	 */
@@ -75,7 +89,7 @@ public class Carte{
 	}
 
 	/**
-	 * Méthode permettant d'enlever un lien à l'index
+	 * Méthode permettant d'enlever un lien à l'index.
 	 * 
 	 * @param index, index où enlever le lien
 	 */
@@ -85,12 +99,15 @@ public class Carte{
 	
 
 	/**
-	 * Méthode permettant d'obtenir une fraction de la liste. La fraction est construite 
-	 * par deep copy.
+	 * Méthode permettant d'obtenir une fraction de la liste. 
+	 * La fraction est construite par deep copy.
 	 * 
-	 * @param duDebut, boolean indiquant si la section du début ou de la fin doit être retourné
-	 * @param indexCoupe, index de coupe
-	 * @return Segment de liste selectionné
+	 * @param duDebut, Valeur booléenne qui indique si la section du début 
+	 *                 doit être retournée.  Autrement, c'est la fin qui 
+	 *                 est retournée. 
+	 *                 
+	 * @param indexCoupe, Index de coupe.
+	 * @return Segment Celui de la liste selectionnée.
 	 */
     public ListeDChainee obtientFraction(boolean duDebut, int indexCoupe){
         
@@ -100,7 +117,7 @@ public class Carte{
         // si du debut
         if(duDebut){
             
-            // copie de 0 à l'index
+            // Copie de 0 à l'index.
             listeLiens.deplacerDebut();
             listeDuplicat.ajouterFin(((Lien)listeLiens.getElement()).copie());
             
@@ -110,7 +127,7 @@ public class Carte{
             
         }else{
             
-            // se place à l'index et copie jusqu'à la fin            
+            // Se place à l'index et copie jusqu'à la fin.            
             listeLiens.deplacerAIndex(indexCoupe);
             listeDuplicat.ajouterFin(((Lien)listeLiens.getElement()).copie());
 
@@ -144,25 +161,27 @@ public class Carte{
 
 
     /**
-     * evalue le score de la carte
+     * Évalue le score de la carte.
      * 
      * @param afficher, true si on veut afficher le score en plus de le calculer
      */
     public void evalueScore(boolean afficher){
     	
     	
-    	// calcul la distance moyenne entre 2 noeuds
+    	// Calcule la distance moyenne entre 2 noeuds.
 		moteurDistanceMoyenne.calculDistanceMoyenne(listeLiens,afficher);
 
 		double distanceMoyenne = moteurDistanceMoyenne.getDistanceMoyenne();
 		
-    	// calcul la longeur totales des liens
+    	// Calcule la longeur totales des liens.
     	double longueur = calculSommeLongueurs();
     	
-    	// calcul le score
-    	score = CONFIGURATION.PENALITE_DISTANCE*distanceMoyenne + 
-    			CONFIGURATION.PENALITE_LONGUEUR*longueur + 
-    			CONFIGURATION.PENALITE_DECONNECTE*moteurDistanceMoyenne.getNbNonConnecte();
+    	// Calcule le score.
+    	score = CONFIGURATION.PENALITE_DISTANCE * distanceMoyenne + 
+    			CONFIGURATION.PENALITE_LONGUEUR * longueur +
+    			
+    			CONFIGURATION.PENALITE_DECONNECTE * 
+    					moteurDistanceMoyenne.getNbNonConnecte();
     	
     }    
     
@@ -188,19 +207,25 @@ public class Carte{
     }
 
     /**
-     * surdéfinition de la méthode permettant d'obtenir un représentation String de l'objet
+     * Redéfinition de la méthode permettant d'obtenir un représentation
+     * String de l'objet.
      * 
-     * return représentation de l'objet sur chaîne de caractères
+     * return La représentation de l'objet en chaîne de caractères.
      */
     public String toString(){
+    	
     	String str = new String();
-    	str += "Liste des liens\n";
+    	str += "Liste des liens\n"
+    			;
     	listeLiens.deplacerDebut();
+    	
     	while(listeLiens.deplacerSuivant()){
     		str += listeLiens.getElement().toString();
     		str += "\n";
     	}
+    	
     	str += "Score: " + score + "\n";
+    	
     	return str;
     }
