+++ src/problemeVilles/PopulationVilles.java
@@ -2,15 +2,25 @@ package problemeVilles;
 
 /**
  * Cette classe est le conteneur de la population de villes. Elle est destiné à
- * être un singleton (aucune mise en oeuvre) et être immuable après l'initialisation.
+ * être un singleton (aucune mise en oeuvre) et être immuable après 
+ * l'initialisation.
  * 
  * Liste des méthodes publiques: 
- *     - PopulationVilles, constructeur qui reçoit le nombre de villes à créer, et coordonnée X et Y max
- *     - getVille, obtient la ville à l'index
- *     - getNbVille, obtient le nombre de villes
- *     - toString, obtient représentation chaine de caractère de la population de villes
+ *     - PopulationVilles, Constructeur qui reçoit le nombre de villes à créer, 
+ *                         et coordonnée X et Y max.
+ *     - getVille, Obtient la ville à l'index.
+ *     - getNbVille, Obtient le nombre de villes.
+ *     - toString, Obtient représentation chaine de caractère de la population 
+ *                 de villes.
  *
  * @author Fred Simard | ETS, 
+ * 
+ * @revision Pierre Bélisle | ETS
+ * 
+ *              principalement : Mettre le code sur 80 colonne max.
+ *                               Aération.
+ *                               Correction orthographique. 
+ * 
  * @version Hiver 2018
  */
 
@@ -22,7 +32,8 @@ public class PopulationVilles{
     private Random rand = new Random();
     
     /**
-     * Constructeur avec paramètres permettant d'initialiser la population de villes
+     * Constructeur avec paramètres permettant d'initialiser la population 
+     * de villes.
      * 
      * @param nbVilles, le nombre de villes à initialiser
      * @param maxX, coordonnée max en X
@@ -37,12 +48,14 @@ public class PopulationVilles{
         for(int i=0; i<nbVilles; i++){
             
         	// initialise la ville (index i)
-            popVilles[i] = new Ville(i, rand.nextDouble()*maxX,rand.nextDouble()*maxY);
+            popVilles[i] = new Ville(i, rand.nextDouble() * maxX,
+            		                    rand.nextDouble() * maxY);
         }
     }
 
     /**
-     * Méthode permettant d'obtenir la ville à l'index donnée
+     * Méthode permettant d'obtenir la ville à l'index donnée.
+     * 
      * @param index, index de la ville à retourner
      * @return référence à la ville à l'index
      */
@@ -51,22 +64,24 @@ public class PopulationVilles{
     }
 
     /**
-     * Méthode permettant d'obtenir le nombre de villes
-     * @return le nombre de villes définit
+     * Méthode permettant d'obtenir le nombre de villes.
+     * 
+     * @return Le nombre de villes définies.
      */
     public int getNbVille(){
         return popVilles.length;
     }
 
     /**
-     * Obtient une représentation String de la population de villes
+     * Obtient une représentation String de la population de villes.
+     * 
      * @return String représentant la population de villse
      */
     public String toString(){
     	String str = new String();
     	str += "Liste des villes\n";
     	
-    	// enchaine la représentation de chaque villes
+    	// Enchaine la représentation de chaque villes.
     	for(int i=0;i<popVilles.length;i++){
     		str += popVilles[i];
     		str += "\n";
